casbin

+++ service/v1/item/item.js
@@ -2,6 +2,7 @@ const { ErrorHandler, statusCodes } = require("../../../helper")
 const sequelize = require("../../../config/db")
 const {QueryTypes} = require("sequelize")
 const  jwt =require('jsonwebtoken')
+const enfo=require("../../../helper/casbin-enforcer")
 const {getItems,postItems, putItems,deleteItems}=require('./sqlqueries')
 const joi=require("joi")
 const userSchema = joi.object({
@@ -33,21 +34,31 @@ class Item {
         try {
           const decoded =await jwt.verify(token, 'secret-key');
           req.user = decoded;
+          return req.user.role;
         } catch (error) {
           res.status(401).send('Invalid token.');
         }
       }
     async getItem(req,res,next){
         try{
-            
-            console.log(req.header)
-            
-            this.authenticateUser(req,res,next)
-            
-            console.log("item services is called");
-            const items = await getItems();
-           // res.render("getItem.ejs",{result:items[0]})
-            res.json(items[0]);
+            let fullUrl =  req.originalUrl;
+            console.log(fullUrl)
+  
+              console.log("item services are called");
+              const role=await this.authenticateUser(req,res,next);
+              const sub = role; // user
+              const obj = fullUrl; // resource
+              const act = req.method; 
+              console.log(sub,obj,act)
+              const enforcer = await enfo;
+              if (await enforcer.enforce(sub, obj, act)) {
+                console.log('Access granted');
+                const items = await getItems();
+                 res.json(items[0]);
+              }
+              else{
+                console.log("Access Denied")
+              }
           } 
         catch (err){
             if(err.statusCodes) throw new ErrorHandler(err.statusCodes, err.message)
@@ -56,17 +67,39 @@ class Item {
     }
     async postItem(req,res,next){
         try{
+          let fullUrl =  req.originalUrl;
+          console.log(fullUrl)
+
             console.log("item services are called");
-            this.authenticateUser(req,res,next);
-            const error=await this.validateUser(req,res,next)
-            if(!error){
-              postItems(req);
-              res.send("item registered")
-            }        
-            else
-            {
-              res.send({ error: error.details[0].message })
-            }           
+            const role=await this.authenticateUser(req,res,next);
+            const sub = role; // user
+            const obj = fullUrl; // resource
+            const act = req.method; 
+            console.log(sub,obj,act)
+            const enforcer = await enfo; 
+            await enforcer.loadPolicy();
+
+            // Retrieve the policy rules from the adapter
+            const policies = await enforcer.getPolicy();
+          
+            // Log the retrieved policies
+            console.log("Policies from the database:");
+            console.log(policies);
+            if (await enforcer.enforce(sub, obj, act)) {
+              console.log('Access granted');
+              const error=await this.validateUser(req,res,next)
+              if(!error){
+                postItems(req);
+                res.send("item registered")
+              }        
+              else
+              {
+                res.send({ error: error.details[0].message })
+              }          
+            } else {
+              console.log('Access denied');
+            }
+   
           }
         catch (err){
             if(err.statusCodes) throw new ErrorHandler(err.statusCodes, err.message)
@@ -76,8 +109,16 @@ class Item {
     async putItem(req, res,next){
       try{
         console.log("put item called")
-        this.authenticateUser(req,res,next)
-        const error=await this.validateUser(req,res,next)
+        let fullUrl =  req.originalUrl;
+        const role=await this.authenticateUser(req,res,next);
+        const sub = role; // user
+        const obj = fullUrl.split("?")[0]; ; // resource
+        const act = req.method
+        console.log(sub,obj,act)
+        const enforcer = await enfo; 
+        if (await enforcer.enforce(sub, obj, act)) {
+          console.log('Access granted');
+          const error=await this.validateUser(req,res,next)
         if(!error){
           putItems(req.query.id,req.body)
           res.send("Item Updated")
@@ -85,7 +126,10 @@ class Item {
         else
         {
           res.send({ error: error.details[0].message })
-        }        
+        }
+      } else {
+        console.log('Access denied');
+      }       
       }
       catch (err){
         if(err.statusCodes) throw new ErrorHandler(err.statusCodes, err.message)
@@ -95,9 +139,21 @@ class Item {
     async deleteItem(req, res,next){
       try{
         console.log("delete item called")
-        this.authenticateUser(req,res,next)
-        deleteItems(req.query.id)
-        res.send("Item deleted")
+        let fullUrl =  req.originalUrl;
+        const role=await this.authenticateUser(req,res,next);
+        const sub = role; // user
+        const obj = fullUrl.split("?")[0]; ; // resource
+        const act = req.method
+        console.log(sub,obj,act)
+        const enforcer = await enfo;
+        if (await enforcer.enforce(sub, obj, act)) {
+          console.log('Access granted'); 
+          deleteItems(req.query.id)
+          res.send("Item deleted")
+        }
+        else {
+          console.log('Access denied');
+        }
       }
       catch (err){
         if(err.statusCodes) throw new ErrorHandler(err.statusCodes, err.message)
